+++ common/src/main/java/jimenezli/neuro21/handler/SoundHandler.java
@@ -1,0 +1,12 @@
+package jimenezli.neuro21.handler;
+
+import jimenezli.neuro21.Neuro21Mod;
+import net.minecraft.resources.ResourceLocation;
+
+import static jimenezli.neuro21.Neuro21Mod.prefix;
+
+public class SoundHandler {
+    public static final ResourceLocation NEUROSAMA_AMBIENT = prefix("neurosama_ambient");
+    public static final ResourceLocation NEUROSAMA_HEART = prefix("neurosama_heart");
+    public static final ResourceLocation EVIL_NEUROSAMA_AMBIENT = prefix("evil_neurosama_ambient");
+}