+++ common/src/main/java/jimenezli/neuro21/util/Neuro21SoundType.java
@@ -1,0 +1,6 @@
+package jimenezli.neuro21.util;
+
+public enum Neuro21SoundType {
+    AMBIENT,
+    HEART
+}