+++ common/src/main/java/jimenezli/neuro21/util/NeurosamaType.java
@@ -1,4 +1,4 @@
-package jimenezli.neuro21.entity;
+package jimenezli.neuro21.util;
 
 public enum NeurosamaType {
     NEUROSAMA,