+++ fabric/src/main/java/jimenezli/neuro21/fabric/Neuro21ModFabric.java
@@ -2,6 +2,7 @@ package jimenezli.neuro21.fabric;
 
 import jimenezli.neuro21.ModEntityTypes;
 import jimenezli.neuro21.ModItems;
+import jimenezli.neuro21.ModSoundEvents;
 import net.fabricmc.api.ModInitializer;
 
 public class Neuro21ModFabric implements ModInitializer {
@@ -9,6 +10,7 @@ public class Neuro21ModFabric implements ModInitializer {
     public void onInitialize() {
         ModEntityTypes.register(null);
         ModEntityTypes.registerAttributes();
+        ModSoundEvents.register(null);
         ModItems.register(null);
     }
 }